% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calculate_wlos_app}
\alias{calculate_wlos_app}
\title{identify and visualises the textual differences between groups}
\usage{
calculate_wlos_app(
  df,
  topic_var,
  text_var = Message,
  top_n = 30,
  filter_by = c("association", "frequency"),
  top_terms_cutoff = 500,
  nrow = 4
)
}
\arguments{
\item{df}{A dataframe.}

\item{topic_var}{The variable that contains the topic label for each post.}

\item{text_var}{The variable containing the text to be compared.}

\item{top_n}{The maximum number of words to include for each topic in the visualisation.}

\item{filter_by}{Whether the top_n terms per category are ordered by their association (high WLO value) or by frequency}

\item{top_terms_cutoff}{The maximum number of terms to select the WLOs from. Only relevant for filter_by = "association"}

\item{nrow}{The number of rows the plots should be displayed with.}
}
\value{
A list containing a table with weighted log-odds calculated for each word in each group, and a visualisation
}
\description{
A function to visualize word frequency and log odds ratio for a given dataset. The function takes several arguments and returns a list containing the visualization and the underlying data.
}
\details{
Plots terms that are more/less likely to appear in conversation across different topics.

When using filter_by = "association" it's important to look at the top_terms_cutoff argument. The higher the value, the more terms you'll include; this can mean you include very low frequency terms. A good starting point is ~ 500-1000. If you wanted to see only the top 50 terms of the whole dataset and how they are distributed across groups, you could use top_terms_cutoff == 50.
}
