% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlierServer}
\alias{outlierServer}
\title{#' Title
#'
#' @param input
#' @param output
#' @param session
#'
#' @return
#' @export
#'
#' @examples}
\usage{
outlierServer(id, df, model, clusters, embedder)
}
\description{
#' Title
#'
#' @param input
#' @param output
#' @param session
#'
#' @return
#' @export
#'
#' @examples
}
