% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reducingParamsUi}
\alias{reducingParamsUi}
\title{Title}
\usage{
reducingParamsUi(id, id_num)
}
\arguments{
\item{id}{parameter for shiny identification}

\item{id_num}{parameter for identification in case where function run multiple times}
}
\description{
Title
}
\keyword{internal}
